requirements.txt
transformers==4.34.1
scipy==1.13.1
scikit-image==0.21.0
scikit-learn==1.2.2
opencv-python==4.10.0.84
Pillow==10.4.0
termcolor==1.1.0
PyYAML==6.0.2
tqdm==4.67.1
absl-py==2.1.0
tensorboard==2.17.0
tensorboardX==2.6.1
PyOpenGL==3.1.0
pyrender==0.1.45
trimesh==3.22.0
click==8.1.7
omegaconf==2.2.3
segmentation_models_pytorch
timm
psutil==5.9.5
lmdb==1.4.1
einops==0.8.1
kornia==0.6.7
gdown==5.2.0
plyfile==1.0.3
natsort
mmcv==1.7.0
pip3 install -U xformers==0.0.22.post3  --index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.1
git+https://github.com/facebookresearch/pytorch3d.git@v0.7.4
matplotlib==3.9.1.post1
mediapipe
open_clip